refactor(studio): split document revision repository
