+++ src/contexts/studio/infrastructure/repository/document.py
@@ -11,17 +11,18 @@ from src.contexts.studio.infrastructure.repository.common import (
     InvalidOperation,
     NotFound,
     Project,
-    RevisionDto,
     Session,
     StudioDatabase,
     _document_dto,
-    _revision_dto,
     _session,
     datetime,
     func,
     new_id,
     select,
     text,
+)
+from src.contexts.studio.infrastructure.repository.document_revisions import (
+    DocumentRevisionRepositoryMixin,
 )
 from src.contexts.studio.infrastructure.repository.document_search import (
     DocumentSearchRepositoryMixin,
@@ -30,7 +31,10 @@ from src.contexts.studio.infrastructure.repository.document_search import (
 __all__ = ["DocumentRepositoryMixin"]
 
 
-class DocumentRepositoryMixin(DocumentSearchRepositoryMixin):
+class DocumentRepositoryMixin(
+    DocumentRevisionRepositoryMixin,
+    DocumentSearchRepositoryMixin,
+):
     database: StudioDatabase
 
     if TYPE_CHECKING:
@@ -212,58 +216,6 @@ class DocumentRepositoryMixin(DocumentSearchRepositoryMixin):
             self._refresh_search(session, document, revision)
             session.flush()
             return _document_dto(document)
-
-    def _current_revision(
-        self,
-        session: Session,
-        document: Document,
-    ) -> DocumentRevision:
-        if document.current_revision_id is None:
-            raise InvalidOperation("Document has no current revision.")
-        revision = session.get(DocumentRevision, document.current_revision_id)
-        if revision is None:
-            raise InvalidOperation("Document revision chain is invalid.")
-        return revision
-
-    def get_revision(
-        self,
-        project_id: str,
-        document_id: str,
-        revision_id: str,
-        *,
-        owner_id: str | None,
-        guest_session_id: str | None,
-    ) -> RevisionDto:
-        with self.database.session() as session:
-            project = self._project(session, project_id, owner_id, guest_session_id)
-            document = self._document(session, project, document_id)
-            revision = session.scalar(
-                select(DocumentRevision).where(
-                    DocumentRevision.id == revision_id,
-                    DocumentRevision.document_id == document.id,
-                )
-            )
-            if revision is None:
-                raise NotFound("Revision not found.")
-            return _revision_dto(revision)
-
-    def list_revisions(
-        self,
-        project_id: str,
-        document_id: str,
-        *,
-        owner_id: str | None,
-        guest_session_id: str | None,
-    ) -> list[RevisionDto]:
-        with self.database.session() as session:
-            project = self._project(session, project_id, owner_id, guest_session_id)
-            document = self._document(session, project, document_id)
-            revisions = session.scalars(
-                select(DocumentRevision)
-                .where(DocumentRevision.document_id == document.id)
-                .order_by(DocumentRevision.revision_number.desc())
-            ).all()
-            return [_revision_dto(revision) for revision in revisions]
 
     def reorder_documents(
         self,
